pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Job Search Telegram Bot with AWS Lambda integration"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.163.0",
    "google-auth>=2.38.0",
    "google-generativeai>=0.8.4",
    "python-telegram-bot>=21.11.1",
    "requests>=2.32.3",
    "boto3>=1.34.0",
]
